cleaning up loans v1 files
