+++ mobile-app/app/screens/AppNavigator/screens/Loans/screens/EditLoanSchemeScreen.tsx
@@ -1,8 +1,7 @@
 import {
-  ThemedScrollView,
-  ThemedSectionTitle,
-  ThemedText,
-  ThemedView,
+  ThemedScrollViewV2,
+  ThemedTextV2,
+  ThemedViewV2,
 } from "@components/themed";
 import {
   LoanScheme,
@@ -12,25 +11,22 @@ import { tailwind } from "@tailwind";
 import { translate } from "@translations";
 import { useEffect, useState } from "react";
 import BigNumber from "bignumber.js";
-import { View } from "@components";
+import { View } from "react-native";
 import { StackScreenProps } from "@react-navigation/stack";
+import { NumericFormat as NumberFormat } from "react-number-format";
 import { useSelector } from "react-redux";
 import { RootState } from "@store";
 import { ascColRatioLoanScheme } from "@store/loans";
-import { Button } from "@components/Button";
 import { hasTxQueued } from "@store/transaction_queue";
 import { hasTxQueued as hasBroadcastQueued } from "@store/ocean";
 import { useWhaleApiClient } from "@shared-contexts/WhaleContext";
 import { useLogger } from "@shared-contexts/NativeLoggingProvider";
-import { getPrecisedTokenValue } from "@screens/AppNavigator/screens/Auctions/helpers/precision-token-value";
-import { useVaultStatus, VaultStatusTag } from "../components/VaultStatusTag";
-import {
-  LoanSchemeOptions,
-  WalletLoanScheme,
-} from "../components/LoanSchemeOptions";
+import { LoanSchemeOptionsV2 } from "@screens/AppNavigator/screens/Loans/components/LoanSchemeOptionsV2";
+import { ButtonV2 } from "@components/ButtonV2";
+import { useVaultStatus } from "../components/VaultStatusTag";
+import { VaultStatus } from "../VaultStatusTypes";
+import { WalletLoanScheme } from "../components/LoanSchemeOptions";
 import { LoanParamList } from "../LoansNavigator";
-import { VaultSectionTextRow } from "../components/VaultSectionTextRow";
-import { useCollateralizationRatioColor } from "../hooks/CollateralizationRatio";
 
 type Props = StackScreenProps<LoanParamList, "EditLoanSchemeScreen">;
 
@@ -61,6 +57,7 @@ export function EditLoanSchemeScreen({
   const hasPendingBroadcastJob = useSelector((state: RootState) =>
     hasBroadcastQueued(state.ocean)
   );
+
   const onSubmit = (): void => {
     if (
       selectedLoanScheme === undefined ||
@@ -126,150 +123,204 @@ export function EditLoanSchemeScreen({
   }
 
   return (
-    <ThemedScrollView contentContainerStyle={tailwind("px-4 pt-0 pb-8")}>
-      <ThemedSectionTitle
-        style={tailwind("text-xs pb-2 pt-4 font-medium")}
-        text={translate(
-          "screens/EditLoanSchemeScreen",
-          "EDIT LOAN SCHEME OF VAULT"
-        )}
-      />
-      <VaultSection vault={activeVault} />
-      <ThemedText
-        light={tailwind("text-gray-500")}
-        dark={tailwind("text-gray-500")}
-        style={tailwind("text-xs mb-6")}
+    <ThemedScrollViewV2 contentContainerStyle={tailwind("px-5 pt-0 pb-8 py-7")}>
+      <VaultDetail vault={activeVault} />
+
+      <ThemedTextV2
+        light={tailwind("text-mono-light-v2-500")}
+        dark={tailwind("text-mono-dark-v2-500")}
+        style={tailwind("font-normal-v2 text-xs mx-5 mt-8")}
       >
-        {translate(
-          "screens/EditLoanSchemeScreen",
-          "Make sure your collateralization ratio is still above your min. collateralization ratio"
-        )}
-      </ThemedText>
-      <LoanSchemeOptions
-        isLoading={!hasFetchedLoanSchemes}
+        {translate("screens/EditLoanSchemeScreen", "AVAILABLE SCHEMES")}
+      </ThemedTextV2>
+      <LoanSchemeOptionsV2
         loanSchemes={filteredLoanSchemes}
+        isLoading={!hasFetchedLoanSchemes}
         selectedLoanScheme={selectedLoanScheme}
         onLoanSchemePress={(scheme: LoanScheme) =>
           setSelectedLoanScheme(scheme)
         }
+        containerStyle={tailwind("mt-2")}
       />
-      <Button
+
+      <View style={tailwind("pt-12 px-12")}>
+        {selectedLoanScheme && (
+          <ThemedTextV2
+            style={tailwind("text-xs font-normal-v2 pb-5 text-center")}
+            light={tailwind("text-mono-light-v2-500")}
+            dark={tailwind("text-mono-dark-v2-500")}
+            testID="action_message"
+          >
+            {translate(
+              "screens/EditLoanSchemeScreen",
+              "Review full details in the next screen"
+            )}
+          </ThemedTextV2>
+        )}
+      </View>
+
+      <ButtonV2
         disabled={
           selectedLoanScheme === undefined ||
           selectedLoanScheme.id === activeVault.loanScheme.id ||
           hasPendingJob ||
           hasPendingBroadcastJob
         }
-        label={translate("screens/EditLoanSchemeScreen", "CONTINUE")}
+        label={translate("screens/EditLoanSchemeScreen", "Continue")}
         onPress={onSubmit}
-        margin="mt-7 mb-2"
+        styleProps="mt-0 mx-7"
         testID="edit_loan_scheme_submit_button"
       />
-      <ThemedText
-        light={tailwind("text-gray-500")}
-        dark={tailwind("text-gray-400")}
-        style={tailwind("text-center text-xs")}
-      >
-        {translate(
-          "screens/EditLoanSchemeScreen",
-          "Confirm your vault details in next screen"
-        )}
-      </ThemedText>
-    </ThemedScrollView>
+    </ThemedScrollViewV2>
   );
 }
 
-function VaultSection(props: { vault: LoanVaultActive }): JSX.Element {
+function VaultDetail(props: { vault: LoanVaultActive }): JSX.Element {
   const { vault } = props;
   const colRatio = new BigNumber(vault.informativeRatio);
   const minColRatio = new BigNumber(vault.loanScheme.minColRatio);
-  const totalLoanValue = new BigNumber(vault.loanValue);
+  const totalLoanAmount = new BigNumber(vault.loanValue);
   const totalCollateralValue = new BigNumber(vault.collateralValue);
+
   const vaultState = useVaultStatus(
     vault.state,
     colRatio,
     minColRatio,
-    totalLoanValue,
+    totalLoanAmount,
     totalCollateralValue
   );
-  const colors = useCollateralizationRatioColor({
-    colRatio,
-    minColRatio,
-    totalLoanAmount: totalLoanValue,
-    totalCollateralValue,
-  });
+
   return (
-    <ThemedView
-      light={tailwind("bg-white border-gray-200")}
-      dark={tailwind("bg-gray-800 border-gray-700")}
-      style={tailwind("border rounded px-4 py-3 mb-2")}
-    >
-      <View style={tailwind("mb-2 flex flex-row")}>
-        <View style={tailwind("flex-1 mr-5")}>
-          <ThemedText
-            style={tailwind("font-medium")}
-            numberOfLines={1}
-            ellipsizeMode="middle"
-          >
-            {vault.vaultId}
-          </ThemedText>
+    <View>
+      <ThemedTextV2
+        light={tailwind("text-mono-light-v2-500")}
+        dark={tailwind("text-mono-dark-v2-500")}
+        style={tailwind("font-normal-v2 text-xs mx-5")}
+      >
+        {translate("screens/EditLoanSchemeScreen", "VAULT DETAILS")}
+      </ThemedTextV2>
+      <ThemedViewV2
+        light={tailwind("bg-mono-light-v2-00")}
+        dark={tailwind("bg-mono-dark-v2-00")}
+        style={tailwind("py-4.5 px-5 rounded-lg-v2 mt-2")}
+      >
+        <View style={tailwind("flex flex-row justify-between")}>
+          <View style={tailwind("w-5/12")}>
+            <ThemedTextV2
+              light={tailwind("text-mono-light-v2-900")}
+              dark={tailwind("text-mono-dark-v2-900")}
+              style={tailwind("font-medium-v2 text-sm")}
+              numberOfLines={1}
+              ellipsizeMode="middle"
+            >
+              {vault.vaultId}
+            </ThemedTextV2>
+            <ThemedTextV2
+              light={tailwind("text-mono-light-v2-700")}
+              dark={tailwind("text-mono-dark-v2-700")}
+              style={tailwind("text-xs")}
+              numberOfLines={1}
+              ellipsizeMode="middle"
+            >
+              {translate(
+                "screens/EditLoanSchemeScreen",
+                "{{interestRate}}% Interest",
+                {
+                  interestRate: vault.loanScheme.interestRate,
+                }
+              )}
+            </ThemedTextV2>
+          </View>
+
+          <View style={tailwind("items-end")}>
+            <NumberFormat
+              displayType="text"
+              renderText={(val: string) => (
+                <ThemedTextV2
+                  light={tailwind(
+                    "text-mono-light-v2-900",
+                    {
+                      "text-green-v2":
+                        vaultState.status === VaultStatus.Ready ||
+                        vaultState.status === VaultStatus.Healthy,
+                    },
+                    {
+                      "text-orange-v2":
+                        vaultState.status === VaultStatus.AtRisk,
+                    },
+                    {
+                      "text-red-v2":
+                        vaultState.status === VaultStatus.NearLiquidation,
+                    }
+                  )}
+                  dark={tailwind(
+                    "text-mono-dark-v2-900",
+                    {
+                      "text-green-v2":
+                        vaultState.status === VaultStatus.Ready ||
+                        vaultState.status === VaultStatus.Healthy,
+                    },
+                    {
+                      "text-orange-v2":
+                        vaultState.status === VaultStatus.AtRisk,
+                    },
+                    {
+                      "text-red-v2":
+                        vaultState.status === VaultStatus.NearLiquidation,
+                    }
+                  )}
+                  style={tailwind("text-sm font-semibold-v2")}
+                >
+                  {vault.informativeRatio === "-1"
+                    ? translate(
+                        "components/VaultCard",
+                        `${vaultState.status
+                          .charAt(0)
+                          .toUpperCase()}${vaultState.status
+                          .slice(1)
+                          .toLowerCase()}`
+                      )
+                    : `${val}%`}
+                </ThemedTextV2>
+              )}
+              thousandSeparator
+              value={new BigNumber(
+                vault.informativeRatio === "-1" ? NaN : vault.informativeRatio
+              ).toFixed(2)}
+            />
+
+            <NumberFormat
+              displayType="text"
+              suffix="%"
+              prefix={translate("screens/EditLoanSchemeScreen", "min. ")}
+              renderText={(value: string) => (
+                <ThemedTextV2
+                  light={tailwind("text-mono-light-v2-700")}
+                  dark={tailwind("text-mono-dark-v2-700")}
+                  style={tailwind("text-xs font-normal-v2")}
+                  testID="text_total_collateral_value"
+                >
+                  {value}
+                </ThemedTextV2>
+              )}
+              thousandSeparator
+              value={new BigNumber(vault.loanScheme.minColRatio ?? 0).toFixed(
+                0
+              )}
+            />
+          </View>
         </View>
-        <VaultStatusTag status={vaultState.status} />
-      </View>
-      <VaultSectionTextRow
-        testID="text_total_collateral_value"
-        value={getPrecisedTokenValue(vault.collateralValue ?? 0)}
-        prefix="$"
-        lhs={translate(
-          "screens/EditCollateralScreen",
-          "Total collateral (USD)"
+      </ThemedViewV2>
+      <ThemedTextV2
+        light={tailwind("text-mono-light-v2-500")}
+        dark={tailwind("text-mono-dark-v2-500")}
+        style={tailwind("font-normal-v2 text-xs mx-5 mt-2")}
+      >
+        {translate(
+          "screens/EditLoanSchemeScreen",
+          "You can only select a scheme lower than your vault's current collateralization."
         )}
-        isOraclePrice
-      />
-      <VaultSectionTextRow
-        testID="text_total_collateral_value"
-        value={getPrecisedTokenValue(vault.loanValue ?? 0)}
-        prefix="$"
-        lhs={translate("screens/EditCollateralScreen", "Total loans (USD)")}
-        isOraclePrice
-      />
-      <VaultSectionTextRow
-        testID="text_total_collateral_value"
-        value={new BigNumber(
-          vault.informativeRatio === "-1" ? NaN : vault.informativeRatio
-        ).toFixed(2)}
-        suffix={
-          vault.informativeRatio === "-1"
-            ? translate("screens/EditCollateralScreen", "N/A")
-            : "%"
-        }
-        suffixType="text"
-        lhs={translate(
-          "screens/EditCollateralScreen",
-          "Collateralization ratio"
-        )}
-        rhsThemedProps={colors}
-        info={{
-          title: "Collateralization ratio",
-          message:
-            "The collateralization ratio represents the amount of collateral deposited in a vault in relation to the loan amount, expressed in percentage.",
-        }}
-      />
-      <VaultSectionTextRow
-        testID="text_total_collateral_value"
-        value={new BigNumber(vault.loanScheme.minColRatio ?? 0).toFixed(2)}
-        suffix="%"
-        suffixType="text"
-        lhs={translate(
-          "screens/EditCollateralScreen",
-          "Min. collateralization ratio"
-        )}
-        info={{
-          title: "Min. collateralization ratio",
-          message:
-            "Minimum required collateralization ratio based on loan scheme selected. A vault will go into liquidation when the collateralization ratio goes below the minimum requirement.",
-        }}
-      />
-    </ThemedView>
+      </ThemedTextV2>
+    </View>
   );
 }
